requirements.txt
# Core Django
Django>=5.1.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# Database
psycopg2-binary>=2.9.9

# CORS handling
django-cors-headers>=4.3.1

# Environment variables
python-decouple>=3.8

# Filtering and search
django-filter>=23.5

# Audit logging
django-auditlog>=2.3.0

# File handling
Pillow>=10.2.0

# API documentation
drf-spectacular>=0.27.0

# Testing
pytest>=7.4.3
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Async tasks (optional for future)
celery>=5.3.4
redis>=5.0.1

# Channels for WebSockets (optional for future)
channels>=4.0.0
channels-redis>=4.1.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
